# Optimise catalog function rebase of #3145