Rebase of https://github.com/distribution/distribution/pull/3145 to sign off commits and unblock v3 release.

Fixes: https://github.com/distribution/distribution/pull/3145

> fixes #3233 OCI Conformance: GET start of tag is set by last query parameter
> 
> Basically this implementation will only stat `n` times, and only look at the dirs required. E.g. with around 6.500 images, this will only call the `rec` function ~100 times (which also means stat that amount), whereas the other will be called ~20.000 times, where it will stat them all.
> 
> This also fixes a bug where if you specify something close to a last entry, e.g. `fo`, and `foo` is meant, then it will work. This should not be the behavior, as it can cause various bugs.
> 
> Removed the `lessPath` as it isn't needed anymore. Fun fact, the non-native method is slower than native.
> 
> ```
> BenchmarkPathCompareEqual-4         	 3906804	         297 ns/op
> BenchmarkPathCompareNotEqual-4      	148567686	         7.93 ns/op
> BenchmarkPathCompareNative-4        	212109276	         5.61 ns/op
> BenchmarkPathCompareNativeEqual-4   	283666616	         3.83 ns/op
> ```
> 
> Should be ready to merge!



---
- To see the specific tasks where the Asana app for GitHub is being used, see below:
  - https://app.asana.com/0/0/1204651419195143