@DavidSpek 2023-06-20
Friendly ping @thaJeztah @squizzi @wy65701436. This PR should finalize OCI conformance and close 3 milestones for the v3 release.