@eyJhb 2023-08-18
Curious, as I can't seem to find it. Was the CodeQL ever fixed? Ie. potential path traversal. I can't seem to find any prevention for this. With was originally why the other PR went stale :)