@milosgajdos 2023-08-18
It was indeed. We wouldnt merge if it werent 😄 Mind you there were also bugs fixed in CodeQL itself, too.

<img width="802" alt="Screenshot 2023-08-18 at 08 58 05" src="https://github.com/distribution/distribution/assets/1392526/a246585b-4245-4158-87f6-5a84101518a9">