@eyJhb 2023-08-18
Hmm, I see no code changes related to it. Maybe CodeQL just got smarter then :p